+++ src/agents/flagAgent.mjs
@@ -1,7 +1,7 @@
 import { promises, readdirSync } from "fs";
 import { fileURLToPath } from "url";
 import path from "path";
-import * as logs from "../modules/logs.mjs";
+import * as logs from "#src/modules/logs.mjs";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const USERS_DIR = path.join(__dirname, "../../userData");
