+++ src/commands/rule.mjs
@@ -1,6 +1,6 @@
 import {MessageFlags, SlashCommandBuilder} from "discord.js";
-import { emojis } from "#src/consts/phantys_home";
-import * as logs from "#src/modules/logs";
+import { emojis } from "#src/consts/phantys_home.mjs";
+import * as logs from "#src/modules/logs.mjs";
 
 import rulesJSON from "#src/consts/rules.json" with { type: "json" };
 const options = rulesJSON.map(object => ({name: object.title, value: object.id})); // Get a list of rules for the rule command
