+++ src/functions/DMFormHandler.mjs
@@ -1,12 +1,12 @@
-import {getClient} from "#src/modules/client";
-import * as database from "#src/modules/database";
-import * as logs from "#src/modules/logs";
-import * as skinForm from "#src/functions/skinFormHandler";
-import {skinTypeFromFooter} from "#src/functions/skinFormHandler";
-import {flags, setFlag} from "#src/agents/flagAgent";
-import {gun_skins} from "#src/consts/gun_skins";
-import {capitalize} from "#src/modules/util";
-import {getMember} from "#src/modules/discord";
+import {getClient} from "#src/modules/client.mjs";
+import * as database from "#src/modules/database.mjs";
+import * as logs from "#src/modules/logs.mjs";
+import * as skinForm from "#src/functions/skinFormHandler.mjs";
+import {skinTypeFromFooter} from "#src/functions/skinFormHandler.mjs";
+import {flags, setFlag} from "#src/agents/flagAgent.mjs";
+import {gun_skins} from "#src/consts/gun_skins.mjs";
+import {capitalize} from "#src/modules/util.mjs";
+import {getMember} from "#src/modules/discord.mjs";
 
 export async function getLastBotMessage(message) {
     let lastMessage;
